test: assertions move to utils
