+++ src/tests/specs/orbiter/orbiter.spec.ts
@@ -608,7 +608,8 @@ describe('Orbiter', () => {
 
 			testControllers({
 				actor: () => actor,
-				controller: () => controller
+				controller: () => controller,
+				pic: () => pic
 			});
 		});
 	});
